Refactors command classes and enum out of single file and into own files
This is a start towards making the modification and extension of these more manageable.
