+++ DEnc/Commands/Enums.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DEnc.Commands
+{
+    internal enum StreamType
+    {
+        Video,
+        Audio,
+        Subtitle,
+        MPD
+    }
+}
